Move trades-to-ohlcv tests
